requirements.txt
base58
pycryptodome
setuptools
requests
urllib3
fastapi
sendgrid~=6.11.0
django-sendgrid-webhook~=0.3.0
passlib~=1.7.4
uvicorn
python-jose
python-multipart
pydantic~=2.6.4
pydantic[email]
uvicorn
authlib~=1.3.0
python-dotenv~=1.0.1
itsdangerous
pytest
pytest-env
httpx
pydantic_core~=2.16.3
starlette~=0.36.3
sqlalchemy
sqlalchemy-utils
psycopg2-binary
asyncpg
alembic
pydantic-settings
yarl
bcrypt
azure-communication-email
anyio~=4.3.0
# uvloop
responses